Checkout: Ensure late renewal fee is applied when upgrading to a plan
If a user on a free plan has a domain that is in redemption, upgrading
to a paid plan did not redeemed the domain. This patch addresses that
(needs D10503-code on the backend).
Additionally, it improves handling of the `meta` field for the plan -
previously, it was the same regardless of the domain that it was bundled
with.