Extract lists of site fields and options to a module with constants
This patch deduplicates the list of site fields and options to fetch into a
`state/sites/constants` modules. Until now, they were duplicated in state/jetpack-connect
and state/sites and slowly getting out of sync. For example, `site_goals` was added only
to one place in #22171.

Also, the list is now an array with one item per line. Makes for a much more readable diffs
when a field is added or removed.